Add contiguity constraint based on geographic distance
Introducing contigConst for contiguity constraint to be used to apply
weighted geographic distances on the initial dissimilarity matrix between
the variable(s) in the input data matrix.

It should be a positive value (recommended between 0 and 1) where
contigConst = 0 means that the clustering is completely based on the data
with no constraints for contiguity while contigConst > 0 gives priority
to merging contiguous points.

Geographic coordinates are required to compute the geographic distances.

Regional lineage method does not support contiguity constraints.